Split code from gtk_webkitgtk.hh
